Statistics files for code that reproduces touchstone LOAR2 run
These are the results that should be found when using CVP.2017.01.17 code.
These are the seaice.stats, ocean.stats and diag_integral.out files from various compilers.

The PGI runs were with SYMMETRIC_MEMORY defined, the Intel runs were without it.

The debug suffix  implies code compile with -O0, the repro with -O2 and openmp with OpenMP enabled.
MOM_override needed to have a line removed for DIAG_REMAP_Z_GRID_DEF
The test runs are for one day for debug, 2 for repro.